legal-agents-teams/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
semantic-kernel==0.4.1.dev0
azure-identity==1.14.1
